Configurable opt-in/opt-out registry policy.
The constructor takes a `bool` to set the registry policy. True for "opt-in" and false for "opt-out". An opt-out registry will force users to specify an implementation to use and lock it in until changed by them. An opt-in registry will serve up the default implementation unless the user explicitly locks in a different implementation version. 

The registry policy cannot be changed because doing so would lead to user surprise.